// 2023-05-01/src/types.ts
export type Category = {
	id: string;
	name: string;
};

export type Image = {
	url: string;
};

export type ProductSummary = {
	id: string;
	category: Category;
	thumbnail: Image;
	name: string;
	price: number;
};

export type ProductOptionItem = {
	id: string;
	name: string;
};

export type ProductOption = {
	id: string;
	name: string;
	items: ProductOptionItem[];
};

export type ProductDetail = {
	id: string;
	category: Category;
	images: Image[];
	name: string;
	price: number;
	options: ProductOption[];
	description: string;
};

export type OrderOptionItem = {
	name: string;
};

export type OrderOption = {
	name: string;
	item: OrderOptionItem;
};

export type LineItem = {
	id: string;
	product: {
		id: string;
		name: string;
	};
	options: OrderOption[];
	unitPrice: number;
	quantity: number;
	totalPrice: number;
};

export type Cart = {
	lineItems: LineItem[];
	totalPrice: number;
};

export const nullProductDetail: ProductDetail = {
	id: '',
	category: {id: '', name: ''},
	images: [],
	name: '',
	price: 0,
	options: [],
	description: '',
};